requirements.txt
Django>=4.0,<5.0
Faker>=18.0.0
psycopg2>=2.9.0
google-generativeai
python-dotenv
plotly>=5.0.0
pandas>=1.3.0
pytest-django>=4.5.2
pytest>=7.0.0
Ipython
openai
anthropic